+++ src/systemTest/resources/system-test-hub.properties
@@ -1,12 +1,12 @@
 helm.chart.path=http://ddt-nexus.flightstats.io/repository/helm/hub.tgz
-#helm.release.name=
+helm.release.name=funky-wildebeest
 #helm.release.delete=false
-hub.url=http://hub-%s.dls.dev.flightstats.io
-callback.url=http://callback-server-%s.dls.dev.flightstats.io
+hub.url=http://hub.iad.dev.flightstats.io
+callback.url=http://callback-server-funky-wildebeest.dls.dev.flightstats.io
 
 #### s3
-s3.url=http://s3-%s.dls.dev.flightstats.io
+s3.url=s3-external-1.amazonaws.com
 s3.region=us-east-1
 s3.credentials.path=/etc/.aws/credentials
-s3.bucket.name=hub_%s-local
+s3.bucket.name=hub-v2-dev-east1.flightstats.com
 s3.mock.enabled=false
